backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
torch
torchvision
segmentation-models-pytorch
albumentations
matplotlib
pillow
